Move Base64 decode to Base64 file
Follows https://github.com/Lithe-Project/Lithe/commit/f5a26dcd865e7f689ec2e5bc4891825df922dc30
