+++ src/Common/Base64.cpp
@@ -1,52 +1,93 @@
 // Copyright (c) 2016-2019, The Karbo developers
-//
-// This file is part of Karbo.
-//
-// Karbo is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Karbo is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Karbo.  If not, see <http://www.gnu.org/licenses/>.
+// Copyright (c) 2017-2018 The Circle Foundation & Conceal Devs
+// Copyright (c) 2018-2019 Conceal Network & Conceal Devs
+
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "Base64.h"
 
-namespace Tools
-{
-  namespace Base64
-  {
-    std::string encode(const std::string& data) {
-      static const char* encodingTable = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-      const size_t resultSize = 4 * ((data.size() + 2) / 3);
-      std::string result;
-      result.reserve(resultSize);
+namespace Tools {
+namespace Base64 {
 
-      for (size_t i = 0; i < data.size(); i += 3) {
-        size_t a = static_cast<size_t>(data[i]);
-        size_t b = i + 1 < data.size() ? static_cast<size_t>(data[i + 1]) : 0;
-        size_t c = i + 2 < data.size() ? static_cast<size_t>(data[i + 2]) : 0;
+namespace {
+  static const std::string base64chars =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+    "abcdefghijklmnopqrstuvwxyz"
+    "0123456789+/";
 
-        result.push_back(encodingTable[a >> 2]);
-        result.push_back(encodingTable[((a & 0x3) << 4) | (b >> 4)]);
-        if (i + 1 < data.size()) {
-          result.push_back(encodingTable[((b & 0xF) << 2) | (c >> 6)]);
-          if (i + 2 < data.size()) {
-            result.push_back(encodingTable[c & 0x3F]);
-          }
-        }
-      }
-
-      while (result.size() != resultSize) {
-        result.push_back('=');
-      }
-
-      return result;
-    }
+  bool is_base64(unsigned char c) {
+    return (isalnum(c) || (c == '+') || (c == '/'));
   }
 }
+
+std::string encode(const std::string& data) {
+  const size_t resultSize = 4 * ((data.size() + 2) / 3);
+  std::string result;
+  result.reserve(resultSize);
+
+  for (size_t i = 0; i < data.size(); i += 3) {
+    size_t a = static_cast<size_t>(data[i]);
+    size_t b = i + 1 < data.size() ? static_cast<size_t>(data[i + 1]) : 0;
+    size_t c = i + 2 < data.size() ? static_cast<size_t>(data[i + 2]) : 0;
+
+    result.push_back(base64chars[a >> 2]);
+    result.push_back(base64chars[((a & 0x3) << 4) | (b >> 4)]);
+    if (i + 1 < data.size()) {
+      result.push_back(base64chars[((b & 0xF) << 2) | (c >> 6)]);
+      if (i + 2 < data.size()) {
+        result.push_back(base64chars[c & 0x3F]);
+      }
+    }
+  }
+
+  while (result.size() != resultSize) {
+    result.push_back('=');
+  }
+
+  return result;
+}
+  
+std::string decode(std::string const& encoded_string) {
+  size_t in_len = encoded_string.size();
+  size_t i = 0;
+  size_t j = 0;
+  size_t in_ = 0;
+  unsigned char char_array_4[4], char_array_3[3];
+  std::string ret;
+
+  while (in_len-- && (encoded_string[in_] != '=') && is_base64(encoded_string[in_])) {
+    char_array_4[i++] = encoded_string[in_]; in_++;
+    if (i == 4) {
+      for (i = 0; i <4; i++)
+        char_array_4[i] = (unsigned char)base64chars.find(char_array_4[i]);
+
+      char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
+      char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
+      char_array_3[2] = ((char_array_4[2] & 0x3) << 6) + char_array_4[3];
+
+      for (i = 0; (i < 3); i++)
+        ret += char_array_3[i];
+      i = 0;
+    }
+  }
+
+  if (i) {
+    for (j = i; j <4; j++)
+      char_array_4[j] = 0;
+
+    for (j = 0; j <4; j++)
+      char_array_4[j] = (unsigned char)base64chars.find(char_array_4[j]);
+
+    char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
+    char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
+    char_array_3[2] = ((char_array_4[2] & 0x3) << 6) + char_array_4[3];
+
+    for (j = 0; (j < i - 1); j++) ret += char_array_3[j];
+  }
+
+  return ret;
+}
+
+} // Base64
+} // Tools
